1983 tests pass, 0 failures.
Added an initialize-heap method to priority queue so that
term-positions-queue could get some control over the size of the heap.
Fixed queue-top to not barf if the queue is empty.  Added a check that
the queue-top is always the minimum entry in the queue, then commented
it out because we have a bug in the multiple-term-doc-pos-enum stuff
that violates that.

Fixed term-position-queue's less-than method to compare documents
correctly.

Added missing doc and freq slots to multiple-term-doc-pos-enum.  Fixed
it's next method to clear the pos-list and set the doc slot correctly.

Added a clone-object method for arrays that preserves adjustability
and the fill-pointer, if any.  Might not be optimal, heh.

Fixed fs-directory's pre-clone protocol initialize-copy method so it
doesn't do the work that clone-object already does.  Did the same for
buffered-index-input, segment-term-enum and term-vectors-reader.

Fixed term-buffer's initialize-copy to just call set-from-term-buffer
like it's supposed to.  Fixed set-from-term-buffer to clone the
text-buf slot instead of copy-seq'ing it (which wasn't preserving
adjustability and the fill-pointer, if any).

Added a print-object method for segment-merge-info, removed an old
FIXME.

Temporarily removed some error checking code in term-infos-io.  It was
failing, but the tests pass without it.  Really need to figure this
out.